# specs/000-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Manager API
  version: 0.3.0
  description: Contrato inicial do MVP. Endpoints autenticados usam sessão em cookie.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
...
paths:
...
  /integrations/whatsapp:
    put:
      tags: [Integrations]
      operationId: configureWhatsApp
      summary: Configurar credenciais WhatsApp do tenant
      x-requirements: [US-001, FR-004, FR-021, BR-008, BR-011]
      parameters:
        - $ref: '#/components/parameters/AntiforgeryToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WhatsAppConfigInput' }
      responses:
        '204': { description: Segredos protegidos e configuração salva }
        '400': { $ref: '#/components/responses/Problem' }
...
  /webhooks/whatsapp:
    get:
      tags: [Webhooks]
      operationId: verifyWhatsAppWebhook
      summary: Verificar challenge do webhook
      security: []
      x-requirements: [FR-005, BR-011]
      parameters:
        - in: query
          name: hub.mode
          required: true
          schema: { type: string }
        - in: query
          name: hub.verify_token
          required: true
          schema: { type: string }
        - in: query
          name: hub.challenge
          required: true
          schema: { type: string }
      responses:
        '200': { description: Challenge validado }
        '403': { description: Token inválido }
    post:
      tags: [Webhooks]
      operationId: receiveWhatsAppWebhook
      summary: Receber evento autenticado da Meta
      security: []
      x-requirements: [US-002, FR-005, FR-006, FR-007, FR-022, BR-011, NFR-001]
      parameters:
        - in: header
          name: X-Hub-Signature-256
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Evento aceito ou já conhecido }
        '401': { description: Assinatura inválida }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-waai-session
  parameters:
    AntiforgeryToken:
      in: header
      name: X-CSRF-TOKEN
      required: true
      schema: { type: string, minLength: 16 }
...
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    WhatsAppConfigInput:
      type: object
      required: [wabaId, phoneNumberId, accessToken]
      properties:
        wabaId: { type: string }
        phoneNumberId: { type: string }
        accessToken: { type: string, writeOnly: true }
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
security:
  - sessionCookie: []